Backend/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.23
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.1
Flask-Migrate==3.1.0
